AI model can now suggest a task and description
