+++ backend/db/connection.js
@@ -16,16 +16,16 @@ export const client = new MongoClient(uri, {
 export const connectTTDB = async () => {
     try {
         await client.connect();
-        console.log("✅ Successfully connected to MongoDB!");
+        console.log("Successfully connected to MongoDB!");
     } catch (error) {
-        console.error("❌ Error connecting to MongoDB:", error);
+        console.error("Error connecting to MongoDB:", error);
     }
 };
 
 // Handle graceful shutdown
 process.on("SIGINT", async () => {
-    console.log("\n🔌 Closing MongoDB connection...");
+    console.log("\nClosing MongoDB connection...");
     await client.close();
-    console.log("🛑 MongoDB connection closed.");
+    console.log("MongoDB connection closed.");
     process.exit(0);
 });
